Complete user story 9 and tests

+++ db/migrate/20210305163727_create_discounts.rb
@@ -1,7 +1,7 @@
 class CreateDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :discounts do |t|
-      t.integer :percentage
+      t.decimal :percentage
       t.integer :quantity
       t.references :merchant, foreign_key: true
       t.timestamps
